removed old unused angular stuff, rename
