+++ public/app/features/plugins/all.ts
@@ -1,7 +1,6 @@
 import './plugin_edit_ctrl';
 import './plugin_page_ctrl';
 import './import_list/import_list';
-import './ds_edit_ctrl';
 import './datasource_srv';
 import './plugin_component';
-import './VariableQueryComponentLoader';
+import './variableQueryEditorLoader';
